+++ database/migrations/2018_03_16_203257_incluye_id_provincia_y_id_localidad_en_persona.php
@@ -29,9 +29,11 @@ class IncluyeIdProvinciaYIdLocalidadEnPersona extends Migration
      */
     public function down()
     {
-        Schema::table('Persona', function (Blueprint $table) {
-            $table->dropColumn(['idProvincia', 'idLocalidad']);
-        });
+        if (Schema::hasColumn('Persona', 'idProvincia')) {
+            Schema::table('Persona', function (Blueprint $table) {
+                $table->dropColumn(['idProvincia', 'idLocalidad']);
+            });
+        }
 
     }
 }
