+++ database/migrations/2018_03_16_203840_incluye_pasaporte_en_persona.php
@@ -29,8 +29,10 @@ class IncluyePasaporteEnPersona extends Migration
      */
     public function down()
     {
-        Schema::table('Persona', function (Blueprint $table) {
-            $table->dropColumn(['pasaporte']);
-        });
+        if (Schema::hasColumn('Persona', 'pasaporte')) {
+            Schema::table('Persona', function (Blueprint $table) {
+                $table->dropColumn(['pasaporte']);
+            });
+        }
     }
 }
