+++ database/migrations/2018_03_20_213534_incluye_campos_de_ubicacion_en_actividad.php
@@ -30,8 +30,10 @@ class IncluyeCamposDeUbicacionEnActividad extends Migration
      */
     public function down()
     {
-        Schema::table('Actividad', function (Blueprint $table) {
-            $table->dropColumn(['idPais', 'idProvincia', 'idLocalidad']);
-        });
+        if (Schema::hasColumn('Actividad', 'idPais')) {
+            Schema::table('Actividad', function (Blueprint $table) {
+                $table->dropColumn(['idPais', 'idProvincia', 'idLocalidad']);
+            });
+        }
     }
 }
